* viame/dev/merge-master: (355 commits)
  Fixed bug that caused ffmpeg reader to fail on some videos
  use keyword override without virtual
  formatting changes from uncrustify
  added missing libraries
  Added a release note.
  Created depth fusion applet
  add newline at end of file
  added release notes
  added function to gather images for depth estimation
  Added a function to find similar cameras by angle
  Update copyright banner in track_oracle (remainder)
  Update copyright banner in track_oracle/example
  Update copyright banner in track_oracle/example
  Update copyright banner in track_oracle/core
  Changes from code review
  Update descriptions for geo_point-type metadata
  Switched buffer source flag to keep reference
  Added the LoggerView from TeleSculptor
  Switched packet back to class variable
  Removed deprecated avcodec_decode_video2 call.
  ...

+++ vital/util/data_stream_reader.cxx
@@ -92,7 +92,7 @@ size_t
 data_stream_reader::
 line_number() const
 {
-  return m_line_count;
+  return static_cast<size_t>(m_line_count);
 }
